% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{isotopedistribution}
\alias{isotopedistribution}
\title{Isotope distribution}
\usage{
isotopedistribution(smiles)
}
\arguments{
\item{smiles}{SMILES notation of the compound}
}
\value{
isotopic abundance coefficient
}
\description{
This function calculates isotopic abundance for a chemical based on SMILES notation. This number can later be used as a coefficient to multiply the area corresponding to
a monoisotopic ion.
}
\examples{
isotopedistribution("CN1C=NC2=C1C(=O)N(C(=O)N2C)C")
}
