% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear regression}
\usage{
linear_regression(y, x, remove_lowest_conc = FALSE)
}
\arguments{
\item{y}{y-values of the data corresponding to x-values}

\item{x}{x-values of the data}
}
\value{
linear regression parameters (slope and intercept) as a list
}
\description{
This function calculates the linear regression parameters from specified x and y values. Additionally, it checks the linearity based on residuals. In case there exists a residual with absolute value higher than 10, the highest value x-y point will be removed and new linear regression is generated without it. At least 5 datapoints have to remain.
}
